Adjust step under `make install`
`make install` was failing and site was unable to launch.

Line 4 of `make install` failing becuase `home/www/.drush` already
exists in the container, and the 2nd part wouldn't execute because
it is chained together using `;` (which says only run command 2
if command 1 succeeds).

Adds `-p` option so `mkdir` will only create the directory if it
doesn't already exist.