Issue #42: Adjust test steps
IMPORTANT! Temporary test content created by Behat is associated
with the "throwaway" user that gets created in the scenario step.
So if we don't want the temporary test content to persist in the
database after testing, we need to "log in" a user role in any
scenario that creates content!

ie. Use "Given I am logged in as a user with the "anonymous user" role"
instead of "Given I am an anonymous user" if createing content in
the scenario. That way it will be deleted when test is done.