Issue #43: Add cron call to build
The Feed created with the Aggregator Module is updated via Cron.
The block displaying the feed will not display if there is no content
in it, so we add a "cron" call to the install script so that the feed
will have content when it builds and will get displayed.