Issue #44: Fix failing test scenario
The scenario that tested the menu navigation wasn't working. Adding
the @javascript tag to the test scenario fixed it!

Although it remains unclear why that would fix it if it doesn't seem
like there is Javascript involved in the Main Menu...also the menu
links still work with javascript turned off in the broswer...